requirements.txt
amqp==5.1.0
amqplib==1.0.2
anyjson==0.2.5
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
black==22.3.0
carrot==0.10.7
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-picklefield==3.0.1
django-rest-swagger==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipdb==0.13.9
ipython==8.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.2.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
